remove newline terminator in echo output
When echoing an empty environment variable, having a newline character
in the echo output would look like a non-empty path string to
ensure-path, and it would output an extra delimiter in its own output.

Strictly speaking, this isn't necessary. I've updated ensure-path to
handle the case of a whitespace-only entry in a path string by dropping
it. This is probably just a better practice in general for scripts, and
reduces risk of other unforseen issues later.